requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangGraph for agent orchestration
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# LLM Providers
google-generativeai>=0.3.0  # Gemini
anthropic>=0.40.0  # Claude (optional)
openai>=1.0.0  # OpenAI and compatible APIs (optional)
httpx>=0.24.0  # Required for custom auth in OpenAI client

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# PDF generation
markdown-pdf>=1.10

# Additional utilities
python-dateutil>=2.8.2
typing-extensions>=4.5.0
pydantic>=2.0.0  # For structured output schemas
